[K/N] Abstract model
